split agent into files
