+++ agent/.gitignore
@@ -1,6 +1,3 @@
-*.ts
-!index.ts
-!character.ts
 .env
 *.env
 .env*
