+++ agent/src/utils.ts
@@ -1,0 +1,3 @@
+export function getSecret(character: Character, secret: string) {
+	return character.settings?.secrets?.[secret] || process.env[secret]
+}
